chore: Update cardServices component styling and add lazy loading for images
