+++ src/app/components/cards/card.css
@@ -50,3 +50,11 @@
 	max-width: 400px;
 	max-height: 200px;
 }
+
+.card {
+	backdrop-filter: blur(16px) saturate(180%);
+	-webkit-backdrop-filter: blur(16px) saturate(180%);
+	background-color: rgba(17, 25, 40, 0.75);
+	border-radius: 12px;
+	border: 1px solid rgba(255, 255, 255, 0.125);
+}
